requirements.txt
# LangGraph and LangChain
langgraph>=0.2.50
langchain>=0.3.12
langchain-anthropic>=0.3.5
langchain-openai>=0.2.14
langchain-core>=0.3.25

# Multi-provider LLM support (Hybrid Routing)
langchain-groq>=0.2.0
langchain-ollama>=0.2.0

# AI/ML
anthropic>=0.40.0
openai>=1.58.1

# Web Framework (for API endpoints if needed)
fastapi>=0.115.6
uvicorn>=0.34.0
pydantic>=2.10.4
pydantic-settings>=2.7.0

# Database
supabase>=2.11.1
psycopg2-binary>=2.9.10
pgvector>=0.3.6

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.1
aiofiles>=24.1.0
httpx-sse

# File Parsing (Context Engine File Ingestion)
pypdf>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.9

# Model Context Protocol
mcp>=1.1.1
httpx-sse>=0.4.0

# Google APIs (for Gmail watch setup)
google-auth>=2.47.0
google-auth-oauthlib>=1.2.4
google-api-python-client>=2.188.0

# Date/Time
python-dateutil>=2.9.0
pytz>=2024.2

# Background Scheduling (Proactive Agent)
APScheduler>=3.10.0

# E2B Sandbox (Secure Code Execution)
e2b-code-interpreter>=1.0.0

# Slack Integration (The Gantry)
slack-bolt>=1.18.0
aiohttp>=3.9.0

# Job Queue - Slow Path processing (ADR-002)
redis>=5.0.0
rq>=1.16.0
